Move default format declaration
